feat: Update folder structure and data loading
The commit updates the folder structure to match the desired layout. It adds new folders for data related to countries, COVID-19 data, and population data. Additionally, it adds a new folder for storing the results of data analysis. The commit also modifies the data loading function to download the OWID COVID-19 data and population data with age directly to the respective folders.